refactor: commentslist moved to the ui folder
