+++ src/widgets/PostList/PostList.tsx
@@ -2,7 +2,7 @@ import styles from "./PostList.module.css"
 import { useCallback, useMemo, useState } from "react"
 import { PostCard } from "../../entities/post/ui/PostCard"
 import { commentsData } from "../../entities/post/model/commentsData.js"
-import { CommentList } from "../CommentList/CommentList.js"
+import { CommentList } from "../CommentList/ui/CommentList.js"
 import { Container } from "../../shared/ui/Container/Container"
 import { PostLengthFilter } from "../../features/PostLengthFilter/ui/PostLengthFilter.js"
 import { filterByLength } from "../../features/PostLengthFilter/lib/filterByLength.js"
